requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
streamlit>=1.29.0
pyttsx3>=2.90
pillow>=10.0.0